Feature - V2 Invite journey (#259)
* Preliminary work on roi step 4

* Finalising work for step 4 of ROI

* [ci skip] AUTOMATED - update fitness functions

* Work on combining the ROI

* [ci skip] AUTOMATED - update fitness functions

* Nearing the end of ROI combining

* Submission serializer now uses submissiondocument_set to filter over submission documents

* [ci skip] AUTOMATED - update fitness functions

* Submissin serializer now marks the return document dictionary as containnig orphans or not.

* [ci skip] AUTOMATED - update fitness functions

* Submission serializer now only returns user-uploaded documents.

Changing the org associated with submission deletes all submission documents

* [ci skip] AUTOMATED - update fitness functions

* black, flake8, and pii

* [ci skip] AUTOMATED - update fitness functions

* Added some docstrings, and now reassigning the organisation related to a submission deletes old references to the previous organisation

* [ci skip] AUTOMATED - update fitness functions

* Creating tests for the new API endpoints

* Added truncated_name to the DocumentSerializer which shortens it if is over 25 characters

* Changed document serializer truncated name length

* Changing org for a ROI now deletes all documents (LOA included).

Child documents are no longer deleted in the Document delete apiview

* Deleting child documents no longer deletes the parent

* Document serializer now dewfaults to empty dict

* Adding missing validation_errors.py

* Started work on the invite journey

* Nearly finished on this journey

* Starting work on inviting representatives

* Adding some validaiton errors.

Invitation serializer now returns the organisation ID and Name without returning a self-referencing org serializer

Same for OrganisationSerializer and invitations

* Adding validation errors

* Added invitation type to the invitation object to differentiate between own org and representative invites.

Added update_status() to Submission obecjt so I can call it from outside the viewset

* Updated validation_errors.py

* Step 4, updated system parameters to include the new email templates for new and existing users.

Updated validation

* black formatting and flake8

* fitness metrics

* fitness functions

* import cleaning

* Optimised imports

* https://sentry.ci.uktrade.digital/organizations/dit/issues/64584/?project=130&query=is%3Aunresolved

* Security group now gets saved to invitation meta so it can be processed when accepting the invite.

* formatting

* Updating tests

* Updated tests

* black formatting

* Fitness functions

* Added tests for OrganisationSerializer and UserSerializer

* CHanged URL name of User viewset hange group action

* Merging invitation acceptance into overall invitation journey (#265)

* Started work on accepting the invite flow. Speeding up serializers and introducing django-ql to speed up requests

* CHanging Invitation model to keep track of the user that's created of the back of it, and also which cases an invitation should grant the user access to

* Invitation processing sand acceptance

* Internationalising phone numbers and finalising journey

* Calls to update a contacts number are automatically formatted to e164 format

* Viewsets now raise a 404 if the object they're retrieving has been marked as deleted

* Removed django-debug-toolbar

* Added django-restql==0.5.3 to requirements

* Added API-wide ability to specify which fields you want the serializer to return based on your POST data, soon to implement the same functionality with query parameters

* Serializers will now ALWAYS return the ID as part of their .data

* Multiple contacts are no longer being created.

Country and postcode are now added to contacts when they are created as part of invite flow

* Changing a lot of the NestedKeyFields to NestedField as per Django-restql.

Added some debug views to raise errors intentionally.

Changed the get_value method of the drf serializer.Serializers object as it was using weird regex to try and find the value in the QueryDict dictionary

* Changed the way the ROI submission is associated with representative contacts and organisations in the backend so it works with V1 caseworker journeys.

Added primary_contact field to Submission to keep track of which contact has been listed as the primary contact in an ROI

* Moved import as causing issues on jenkins

* Organisation name is now blank to avoid serialisation errors on PUT requests

Added ID to TwoFactorAuth serializer

Added validated to organisation serializer

* black and PII

* Changed tests

* Serializer errors are now logged to sentry

* black

+++ trade_remedies_api/core/services/api.py
@@ -363,9 +363,9 @@ class PublicUserApiView(TradeRemediesApiView):
                 raise NotFoundApiExceptions("Invalid parameters of access denied")
         request_data = request.data.dict()
         if invitation:
-            request_data["group"] = invitation.meta["group"]
-            request_data["active"] = invitation.meta["is_active"]
-            request_data["case_spec"] = invitation.meta["case_spec"]
+            request_data["group"] = invitation.meta.get("group")
+            request_data["active"] = invitation.meta.get("is_active")
+            request_data["case_spec"] = invitation.meta.get("case_spec")
             request_data["email"] = invitation.email
             request_data["contact"] = invitation.contact
             organisation = invitation.organisation
@@ -664,7 +664,8 @@ class CreatePendingUserAPI(TradeRemediesApiView):
             invite = Invitation.objects.get(id=invitation_id, organisation=organisation_id)
             invite.set_user_context(request.user)
             if not invite.accepted_at:
-                invite.contact.delete()
+                if invite.contact:
+                    invite.contact.delete()
                 invite.delete()
             else:
                 raise Invitation.DoesNotExist()
